Fixing Controllers
The Controllers were having an issue so I change the last lines where I change Btn6 to Btn6Xmtr2 with a faster motor speed and same with Btn5 to Btn5Xmtr2. I also commented out //while (true) {} // Let's stop here... I'd rather not go through the trouble of doing anything  else. because it was interfering with the code.